requirements.txt
Django==6.0.3
djangorestframework==3.17.1
django-filter==25.2
Pillow==12.1.1
python-dotenv==1.2.2
